Delete license EULA
An end-user license agreement (short EULA) is a legal contract entered into between a software developer or vendor and the user of the software, often where the software has been purchased by the user from an intermediary such as a retailer. A EULA specifies in detail the rights and restrictions which apply to the use of the software.[1]

Form contracts for digital services (such as terms of service and privacy policies) were traditionally presented on paper (see shrink-wrap agreement) but are now often presented digitally via browsewrap or clickwrap[2][3] formats. As the user may not see the agreement until after they have already purchased or engaged with the software, these documents may be contracts of adhesion.

Software companies often make special agreements with large businesses and government entities that include support contracts and specially drafted warranties.

Many EULAs assert extensive liability limitations. Most commonly, an EULA will attempt to hold harmless the software licensor in the event that the software causes damage to the user's computer or data, but some software also proposes limitations on whether the licensor can be held liable for damage that arises through improper use of the software (for example, incorrectly using tax preparation software and incurring penalties as a result). One case upholding such limitations on consequential damages is M.A. Mortenson Co. v. Timberline Software Corp., et al.[4] Some EULAs also claim restrictions on venue and applicable law in the event that a legal dispute arises.

Some copyright owners use EULAs in an effort to circumvent limitations the applicable copyright law places on their copyrights (such as the limitations in sections 107–122 of the United States Copyright Act), or to expand the scope of control over the work into areas for which copyright protection is denied by law (such as attempting to charge for, regulate or prevent private performances of a work beyond a certain number of performances or beyond a certain period of time). Such EULAs are, in essence, efforts to gain control, by contract, over matters upon which copyright law precludes control.[5] This kind of EULAs concurs in aim with DRM and both may be used as alternate methods for widening control over software.

In disputes of this nature in the United States, cases are often appealed and different circuit courts of appeal sometimes disagree about these clauses. This provides an opportunity for the U.S. Supreme Court to intervene, which it has usually done in a scope-limited and cautious manner, providing little in the way of precedent or settled law.[citation needed]

End-user license agreements are usually lengthy, and written in highly specific legal language, making it difficult for the average user to give informed consent.[6] If the company designs the end-user license agreement in a way that intentionally discourages users from reading them, and uses difficult to understand language, many of the users may not be giving informed consent.